+++ Modules/Monitoring/Providers/MonitoringServiceProvider.php
@@ -16,6 +16,7 @@ use Modules\Monitoring\Livewire\MonitoringMonitorsIndex;
 use Modules\Monitoring\Livewire\MonitoringServerMetricsIndex;
 use Modules\Monitoring\Livewire\MonitoringPreferencesIndex;
 use Modules\Monitoring\Livewire\MonitoringServersIndex;
+use Modules\Monitoring\Livewire\MonitoringServerShow;
 use Modules\Monitoring\Livewire\MonitoringStatusPageSettings;
 use Modules\Monitoring\Livewire\PublicStatusPage;
 
@@ -32,6 +33,7 @@ class MonitoringServiceProvider extends ServiceProvider
         Livewire::component('monitoring.hub-index', MonitoringHubIndex::class);
         Livewire::component('monitoring.monitoring-index', MonitoringIndex::class);
         Livewire::component('monitoring.servers-index', MonitoringServersIndex::class);
+        Livewire::component('monitoring.server-show', MonitoringServerShow::class);
         Livewire::component('monitoring.monitors-index', MonitoringMonitorsIndex::class);
         Livewire::component('monitoring.monitor-show', MonitoringMonitorShow::class);
         Livewire::component('monitoring.server-metrics-index', MonitoringServerMetricsIndex::class);